backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib
python-multipart
pydantic
python-socketio
nltk
python-dotenv
bcrypt
email-validator==2.2.0
openai
pyttsx3
SpeechRecognition
vosk
chromadb
pyaudio
langchain
langchain-community
langchain-openai
redlines
httpx
python-Levenshtein